new: add git todo <number> commit
This is the first commit done with that command. :) The original text of
the todo is as follows:

consider a workflow where todos are commit messages of commits to be
done. Maybe add a new command such as commit and prefix the line with a '*'
instead of a '-' to differentiate them, and limit the message length
to 70 as for the commit messages. Then by reordering them in the file,
you can prepare the order in which you can to do the commits (i.e.
ensure the todos are ordered by file line and not by date). Or simply
add a new command such as 'git todo <number> commit' to use that
message and automatically mark it as done afterwards. :) If so, add a
new status called 'c' to indicate at the end '(Done and committed on
<date>)'. If so, change the status for 'cancelled' e.g. with a '-'.